Fazendo a ordenação na tabela usando query. Acertando o código que fazia o TerraView cair quando fazia promote da tabela.

+++ src/terralib/qt/widgets/table/DataSetTableModel.cpp
@@ -203,7 +203,7 @@ te::da::ObjectIdSet* te::qt::widgets::DataSetTableModel::getObjectIdSet (const i
 
   for(int i=initRow; i<=finalRow; i++)
   {
-    row = (m_promoter == 0) ? i : m_promoter->getLogicalRow(i);
+    row = (m_promoter == 0) ? i : (int)m_promoter->getLogicalRow(i);
     m_dataset->move(row);
 
     oids->add(te::da::GenerateOID(m_dataset, oids->getPropertyNames()));
